Put token functionality in custom file
