+++ resources/js/cache-evader.js
@@ -1,43 +1,7 @@
-import { get as getCookie } from 'browser-cookies';
+import attachTokens from './lib/attach-tokens';
 
-const handleForm = (doc, tokenInput, xsrfToken) => {
-  const xsrfInput = doc.createElement('input');
-
-  xsrfInput.setAttribute('type', 'hidden');
-  xsrfInput.setAttribute('name', '_xsrf_token');
-  xsrfInput.setAttribute('value', xsrfToken);
-
-  const parent = tokenInput.parentNode;
-
-  parent.insertBefore(xsrfInput, tokenInput);
-  parent.removeChild(tokenInput);
-};
-
-const getXsrfToken = async () => {
-  const xsrfToken = getCookie('XSRF-TOKEN');
-
-  if (xsrfToken) {
-    return xsrfToken;
-  }
-
-  await fetch('/cache-evader/ping', {
-    method: 'GET',
-    credentials: 'same-origin',
-  });
-
-  return getCookie('XSRF-TOKEN');
-}
-
-(async function(win) {
+(function(win) {
   const doc = win.document;
 
-  const tokenInputs = [...doc.querySelectorAll('input[name="_token"]')];
-
-  if (!tokenInputs.length) {
-    return;
-  }
-
-  const xsrfToken = await getXsrfToken();
-
-  tokenInputs.forEach(tokenInput => handleForm(doc, tokenInput, xsrfToken));
+  attachTokens(win, doc);
 })(global);
